set multiple and merge into button_group
