+++ crates/ui/src/lib.rs
@@ -39,7 +39,6 @@ pub mod switch;
 pub mod tab;
 pub mod table;
 pub mod theme;
-pub mod toggle_button;
 pub mod tooltip;
 pub mod webview;
